+++ src/datasource-zabbix/utils.ts
@@ -1,5 +1,16 @@
 import _ from 'lodash';
 import moment from 'moment';
+import kbn from 'grafana/app/core/utils/kbn';
+import * as c from './constants';
+import { VariableQuery, VariableQueryTypes } from './types';
+
+/*
+ * This regex matches 3 types of variable reference with an optional format specifier
+ * \$(\w+)                          $var1
+ * \[\[([\s\S]+?)(?::(\w+))?\]\]    [[var2]] or [[var2:fmt2]]
+ * \${(\w+)(?::(\w+))?}             ${var3} or ${var3:fmt3}
+ */
+export const variableRegex = /\$(\w+)|\[\[([\s\S]+?)(?::(\w+))?\]\]|\${(\w+)(?:\.([^:^\}]+))?(?::(\w+))?}/g;
 
 /**
  * Expand Zabbix item name
@@ -12,8 +23,8 @@ export function expandItemName(name, key) {
 
   // extract params from key:
   // "system.cpu.util[,system,avg1]" --> ["", "system", "avg1"]
-  let key_params_str = key.substring(key.indexOf('[') + 1, key.lastIndexOf(']'));
-  let key_params = splitKeyParams(key_params_str);
+  const key_params_str = key.substring(key.indexOf('[') + 1, key.lastIndexOf(']'));
+  const key_params = splitKeyParams(key_params_str);
 
   // replace item parameters
   for (let i = key_params.length; i >= 1; i--) {
@@ -32,10 +43,10 @@ export function expandItems(items) {
 }
 
 function splitKeyParams(paramStr) {
-  let params = [];
+  const params = [];
   let quoted = false;
   let in_array = false;
-  let split_symbol = ',';
+  const split_symbol = ',';
   let param = '';
 
   _.forEach(paramStr, symbol => {
@@ -69,9 +80,9 @@ export function containsMacro(itemName) {
 
 export function replaceMacro(item, macros, isTriggerItem) {
   let itemName = isTriggerItem ? item.url : item.name;
-  let item_macros = itemName.match(MACRO_PATTERN);
+  const item_macros = itemName.match(MACRO_PATTERN);
   _.forEach(item_macros, macro => {
-    let host_macros = _.filter(macros, m => {
+    const host_macros = _.filter(macros, m => {
       if (m.hostid) {
         if (isTriggerItem) {
           // Trigger item can have multiple hosts
@@ -93,10 +104,10 @@ export function replaceMacro(item, macros, isTriggerItem) {
       }
     });
 
-    let macro_def = _.find(host_macros, { macro: macro });
+    const macro_def = _.find(host_macros, { macro: macro });
     if (macro_def && macro_def.value) {
-      let macro_value = macro_def.value;
-      let macro_regex = new RegExp(escapeMacro(macro));
+      const macro_value = macro_def.value;
+      const macro_regex = new RegExp(escapeMacro(macro));
       itemName = itemName.replace(macro_regex, macro_value);
     }
   });
@@ -109,17 +120,62 @@ function escapeMacro(macro) {
   return macro;
 }
 
+export function parseLegacyVariableQuery(query: string): VariableQuery {
+  let queryType: VariableQueryTypes;
+  const parts = [];
+
+  // Split query. Query structure: group.host.app.item
+  _.each(splitTemplateQuery(query), part => {
+    // Replace wildcard to regex
+    if (part === '*') {
+      part = '/.*/';
+    }
+    parts.push(part);
+  });
+  const template = _.zipObject(['group', 'host', 'app', 'item'], parts);
+
+  if (parts.length === 4 && template.app === '/.*/') {
+    // Search for all items, even it's not belong to any application
+    template.app = '';
+  }
+
+  switch (parts.length) {
+    case 1:
+      queryType = VariableQueryTypes.Group;
+      break;
+    case 2:
+      queryType = VariableQueryTypes.Host;
+      break;
+    case 3:
+      queryType = VariableQueryTypes.Application;
+      break;
+    case 4:
+      queryType = VariableQueryTypes.Item;
+      break;
+  }
+
+  const variableQuery: VariableQuery = {
+    queryType,
+    group: template.group || '',
+    host: template.host || '',
+    application: template.app || '',
+    item: template.item || '',
+  };
+
+  return variableQuery;
+}
+
 /**
  * Split template query to parts of zabbix entities
  * group.host.app.item -> [group, host, app, item]
  * {group}{host.com} -> [group, host.com]
  */
 export function splitTemplateQuery(query) {
-  let splitPattern = /\{[^\{\}]*\}|\{\/.*\/\}/g;
+  const splitPattern = /\{[^\{\}]*\}|\{\/.*\/\}/g;
   let split;
 
   if (isContainsBraces(query)) {
-    let result = query.match(splitPattern);
+    const result = query.match(splitPattern);
     split = _.map(result, part => {
       return _.trim(part, '{}');
     });
@@ -131,7 +187,7 @@ export function splitTemplateQuery(query) {
 }
 
 function isContainsBraces(query) {
-  let bracesPattern = /^\{.+\}$/;
+  const bracesPattern = /^\{.+\}$/;
   return bracesPattern.test(query);
 }
 
@@ -143,9 +199,9 @@ export function isRegex(str) {
 }
 
 export function isTemplateVariable(str, templateVariables) {
-  var variablePattern = /^\$\w+/;
+  const variablePattern = /^\$\w+/;
   if (variablePattern.test(str)) {
-    var variables = _.map(templateVariables, variable => {
+    const variables = _.map(templateVariables, variable => {
       return '$' + variable.name;
     });
     return _.includes(variables, str);
@@ -154,10 +210,22 @@ export function isTemplateVariable(str, templateVariables) {
   }
 }
 
+export function getRangeScopedVars(range) {
+  const msRange = range.to.diff(range.from);
+  const sRange = Math.round(msRange / 1000);
+  const regularRange = kbn.secondsToHms(msRange / 1000);
+  return {
+    __range_ms: { text: msRange, value: msRange },
+    __range_s: { text: sRange, value: sRange },
+    __range: { text: regularRange, value: regularRange },
+    __range_series: {text: c.RANGE_VARIABLE_VALUE, value: c.RANGE_VARIABLE_VALUE},
+  };
+}
+
 export function buildRegex(str) {
-  var matches = str.match(regexPattern);
-  var pattern = matches[1];
-  var flags = matches[2] !== "" ? matches[2] : undefined;
+  const matches = str.match(regexPattern);
+  const pattern = matches[1];
+  const flags = matches[2] !== "" ? matches[2] : undefined;
   return new RegExp(pattern, flags);
 }
 
@@ -168,18 +236,18 @@ export function escapeRegex(value) {
 }
 
 export function parseInterval(interval) {
-  var intervalPattern = /(^[\d]+)(y|M|w|d|h|m|s)/g;
-  var momentInterval = intervalPattern.exec(interval);
+  const intervalPattern = /(^[\d]+)(y|M|w|d|h|m|s)/g;
+  const momentInterval: any[] = intervalPattern.exec(interval);
   return moment.duration(Number(momentInterval[1]), momentInterval[2]).valueOf();
 }
 
 export function parseTimeShiftInterval(interval) {
-  let intervalPattern = /^([\+\-]*)([\d]+)(y|M|w|d|h|m|s)/g;
-  let momentInterval = intervalPattern.exec(interval);
-  let duration = 0;
+  const intervalPattern = /^([\+\-]*)([\d]+)(y|M|w|d|h|m|s)/g;
+  const momentInterval: any[] = intervalPattern.exec(interval);
+  let duration: any = 0;
 
   if (momentInterval[1] === '+') {
-    duration = 0 - moment.duration(Number(momentInterval[2]), momentInterval[3]).valueOf();
+    duration = 0 - (moment.duration(Number(momentInterval[2]), momentInterval[3]).valueOf() as any);
   } else {
     duration = moment.duration(Number(momentInterval[2]), momentInterval[3]).valueOf();
   }
@@ -195,13 +263,13 @@ export function parseTimeShiftInterval(interval) {
  */
 export function formatAcknowledges(acknowledges) {
   if (acknowledges.length) {
-    var formatted_acknowledges = '<br><br>Acknowledges:<br><table><tr><td><b>Time</b></td>'
+    let formatted_acknowledges = '<br><br>Acknowledges:<br><table><tr><td><b>Time</b></td>'
       + '<td><b>User</b></td><td><b>Comments</b></td></tr>';
-    _.each(_.map(acknowledges, function (ack) {
-      var timestamp = moment.unix(ack.clock);
+    _.each(_.map(acknowledges, ack => {
+      const timestamp = moment.unix(ack.clock);
       return '<tr><td><i>' + timestamp.format("DD MMM YYYY HH:mm:ss") + '</i></td><td>' + ack.alias
         + ' (' + ack.name + ' ' + ack.surname + ')' + '</td><td>' + ack.message + '</td></tr>';
-    }), function (ack) {
+    }), ack => {
       formatted_acknowledges = formatted_acknowledges.concat(ack);
     });
     formatted_acknowledges = formatted_acknowledges.concat('</table>');
@@ -212,8 +280,8 @@ export function formatAcknowledges(acknowledges) {
 }
 
 export function convertToZabbixAPIUrl(url) {
-  var zabbixAPIUrlPattern = /.*api_jsonrpc.php$/;
-  var trimSlashPattern = /(.*?)[\/]*$/;
+  const zabbixAPIUrlPattern = /.*api_jsonrpc.php$/;
+  const trimSlashPattern = /(.*?)[\/]*$/;
   if (url.match(zabbixAPIUrlPattern)) {
     return url;
   } else {
@@ -246,7 +314,7 @@ export function callOnce(func, promiseKeeper) {
  */
 export function sequence(funcsArray) {
   return function(result) {
-    for (var i = 0; i < funcsArray.length; i++) {
+    for (let i = 0; i < funcsArray.length; i++) {
       result = funcsArray[i].call(this, result);
     }
     return result;
@@ -291,5 +359,5 @@ export function getArrayDepth(a, level = 0) {
 
 // Fix for backward compatibility with lodash 2.4
 if (!_.includes) {
-  _.includes = _.contains;
+  _.includes = (_ as any).contains;
 }