+++ src/panel-triggers/specs/migrations.spec.ts
@@ -4,17 +4,22 @@ import {TriggerPanelCtrl} from '../triggers_panel_ctrl';
 import {DEFAULT_TARGET, DEFAULT_SEVERITY, PANEL_DEFAULTS} from '../triggers_panel_ctrl';
 import {CURRENT_SCHEMA_VERSION} from '../migrations';
 
-describe('Triggers Panel schema migration', () => {
-  let ctx = {};
-  let updatePanelCtrl;
-  let datasourceSrvMock = {
-    getMetricSources: () => {
-      return [{ meta: {id: 'alexanderzobnin-zabbix-datasource'}, value: {}, name: 'zabbix_default' }];
-    },
-    get: () => Promise.resolve({})
+jest.mock('@grafana/runtime', () => {
+  return {
+    getDataSourceSrv: () => ({
+      getMetricSources: () => {
+        return [{ meta: {id: 'alexanderzobnin-zabbix-datasource'}, value: {}, name: 'zabbix_default' }];
+      },
+      get: () => Promise.resolve({})
+    }),
   };
+}, {virtual: true});
 
-  let timeoutMock = () => {};
+describe('Triggers Panel schema migration', () => {
+  let ctx: any = {};
+  let updatePanelCtrl;
+
+  const timeoutMock = () => {};
 
   beforeEach(() => {
     ctx = {
@@ -43,18 +48,20 @@ describe('Triggers Panel schema migration', () => {
       }
     };
 
-    updatePanelCtrl = (scope) => new TriggerPanelCtrl(scope, {}, timeoutMock, datasourceSrvMock, {}, {}, {}, mocks.timeSrvMock);
+    updatePanelCtrl = (scope) => new TriggerPanelCtrl(scope, {}, timeoutMock, {}, {}, {}, mocks.timeSrvMock);
   });
 
   it('should update old panel schema', () => {
-    let updatedPanelCtrl = updatePanelCtrl(ctx.scope);
+    const updatedPanelCtrl = updatePanelCtrl(ctx.scope);
 
-    let expected = _.defaultsDeep({
+    const expected = _.defaultsDeep({
       schemaVersion: CURRENT_SCHEMA_VERSION,
-      datasources: ['zabbix'],
-      targets: {
-        'zabbix': DEFAULT_TARGET
-      },
+      targets: [
+        {
+          ...DEFAULT_TARGET,
+          datasource: 'zabbix',
+        }
+      ],
       ageField: true,
       statusField: false,
       severityField: false,
@@ -68,29 +75,29 @@ describe('Triggers Panel schema migration', () => {
 
   it('should create new panel with default schema', () => {
     ctx.scope.panel = {};
-    let updatedPanelCtrl = updatePanelCtrl(ctx.scope);
+    const updatedPanelCtrl = updatePanelCtrl(ctx.scope);
 
-    let expected = _.defaultsDeep({
+    const expected = _.defaultsDeep({
       schemaVersion: CURRENT_SCHEMA_VERSION,
-      datasources: ['zabbix_default'],
-      targets: {
-        'zabbix_default': DEFAULT_TARGET
-      }
+      targets: [{
+        ...DEFAULT_TARGET,
+        datasource: 'zabbix_default'
+      }]
     }, PANEL_DEFAULTS);
     expect(updatedPanelCtrl.panel).toEqual(expected);
   });
 
   it('should set default targets for new panel with empty targets', () => {
     ctx.scope.panel = {
-      targets: [{}]
+      targets: []
     };
-    let updatedPanelCtrl = updatePanelCtrl(ctx.scope);
+    const updatedPanelCtrl = updatePanelCtrl(ctx.scope);
 
-    let expected = _.defaultsDeep({
-      datasources: ['zabbix_default'],
-      targets: {
-        'zabbix_default': DEFAULT_TARGET
-      },
+    const expected = _.defaultsDeep({
+      targets: [{
+        ...DEFAULT_TARGET,
+        datasource: 'zabbix_default'
+      }]
     }, PANEL_DEFAULTS);
 
     expect(updatedPanelCtrl.panel).toEqual(expected);