+++ src/test-setup/jest-setup.js
@@ -45,6 +45,7 @@ jest.mock('grafana/app/core/utils/datemath', () => {
 jest.mock('grafana/app/core/utils/kbn', () => {
   return {
     round_interval: n => n,
+    secondsToHms: n => n + 'ms'
   };
 }, {virtual: true});
 
@@ -74,9 +75,9 @@ jest.mock('grafana/app/core/config', () => {
 
 jest.mock('jquery', () => 'module not found', {virtual: true});
 
-jest.mock('@grafana/ui', () => {
-  return {};
-}, {virtual: true});
+// jest.mock('@grafana/ui', () => {
+//   return {};
+// }, {virtual: true});
 
 // Required for loading angularjs
 let dom = new JSDOM('<html><head><script></script></head><body></body></html>');